서버,클라 간 룸 단위 시간 동기화

+++ Project_Animagus/Source/Project_Animagus/UI/MyPlayerHUDWidget.h
@@ -37,7 +37,7 @@ public:
     void UpdateCountdown(float CountdownValue);
 
     UFUNCTION(BlueprintCallable, Category="UI")
-    void UpdateRoundTime(float RoundTimeValue);
+    void UpdateRoundTime(int64 RoundTimeValue);
 
     UFUNCTION(BlueprintCallable, Category="UI")
     void UpdateSelectedSkillOutline(int32 SelectedIndex);
